backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
pydantic
python-dotenv

